[driver] Try to prevent spurious routing entries for iSCSI
If a gateway is specified in the iBFT, the Microsoft iSCSI initiator
will create a static route to the iSCSI target via that gateway.  (See
http://support.microsoft.com/kb/960104 for details.)  If the target is
in the same subnet then this is undesirable, since it will mean
duplicating every outbound packet on the network.

Try to prevent this undesirable behaviour by adjusting the gateway
address stored in the iBFT.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>